refactor(avatar): add src, tests, and types folders
