+++ packages/avatar/index.d.ts
@@ -1,2 +1,2 @@
-export { default } from './Avatar';
-export * from './Avatar';
+export { default } from './types/Avatar';
+export * from './types/Avatar';
