+++ packages/avatar/index.js
@@ -1,3 +1,1 @@
-import Avatar from './Avatar';
-
-export default Avatar;
+export { default } from './src/Avatar';
