+++ packages/avatar/package.json
@@ -4,6 +4,7 @@
   "author": "Tyson Warner <tyson.warner@availity.com>",
   "description": "Availity User avatar component",
   "main": "index.js",
+  "types": "index.d.ts",
   "keywords": [
     "react",
     "availity",
@@ -23,22 +24,20 @@
   "devDependencies": {
     "@availity/api-axios": "^5.5.14",
     "@availity/api-core": "^7.0.2",
-    "@availity/localstorage-core": "^3.0.0",
     "axios": "^0.21.1",
-    "prop-types": "^15.7.2",
     "react": "^17.0.2",
     "react-dom": "^17.0.2"
   },
   "peerDependencies": {
     "@availity/api-axios": "^5.4.0",
     "@availity/api-core": "^5.4.0",
-    "@availity/localstorage-core": "^2.6.1",
-    "axios": "^0.16.0 || ^0.17.0 || ^0.18.0 || ^0.19.0",
+    "axios": "^0.21.1",
     "react": ">=16.8.0"
   },
   "dependencies": {
     "@availity/hooks": "^3.0.2",
     "lodash": "^4.17.21",
+    "prop-types": "^15.7.2",
     "react-image": "^4.0.3",
     "react-loading-skeleton": "^2.2.0"
   }
