+++ packages/avatar/tests/Avatar.test.js
