+++ packages/avatar/types/Avatar.d.ts
@@ -1,0 +1,13 @@
+type SkeletonType = {
+  width?: string | number;
+  height?: string | number;
+};
+
+export interface AvatarProps {
+  fallback?: string;
+  skeletonProps?: SkeletonType;
+}
+
+declare const Avatar: React.FC<AvatarProps>;
+
+export default Avatar;
